+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## Unreleased #unstable
 
+ * better diagnostics for the recursive operators, see #272 
  * Use a staged docker build, reducing container size ~70%, see #195
  * Use [Z3-TurnKey](https://github.com/tudo-aqua/z3-turnkey) instead of a
    bespoke Z3 build, see #219